Rename Zapper GraphQL payload generic
