+++ Packages/PortuNetwork/Tests/PortuNetworkTests/ZapperProviderTests.swift
@@ -6,6 +6,7 @@ import Testing
 // MARK: - Tests
 
 @Suite(.serialized)
+// swiftlint:disable:next type_body_length
 struct ZapperProviderTests {
     let session = makeMockSession()
 
@@ -182,6 +183,24 @@ struct ZapperProviderTests {
     }
 
     @Test
+    func `fetchBalances skips null partial edges and nodes`() async throws {
+        defer { ZapperMockURLProtocol.reset() }
+        ZapperMockURLProtocol.requestHandler = { _ in
+            try (jsonData(tokenResponse(edges: [
+                NSNull(),
+                ["node": NSNull()],
+                tokenBalanceEdge(symbol: "ETH"),
+                tokenBalanceEdge(symbol: "DAI", tokenAddress: "0xdai")
+            ])), 200)
+        }
+
+        let provider = makeProvider(session: session)
+        let results = try await provider.fetchBalances(context: makeSyncContext())
+
+        #expect(results.map { $0.tokens[0].symbol } == ["ETH", "DAI"])
+    }
+
+    @Test
     func `graphql auth and rate limit codes map to provider errors`() async throws {
         let cases: [(String, (ZapperError) -> Bool)] = [
             ("UNAUTHENTICATED", { if case .unauthorized = $0 { true } else { false } }),
@@ -320,6 +339,34 @@ struct ZapperProviderTests {
 
         #expect(position.positionType == .staking)
         #expect(position.tokens.first?.role == .lpToken)
+    }
+
+    @Test
+    func `fetchDeFiPositions skips null partial app edges position edges and tokens`() async throws {
+        let validContractEdge = contractPositionEdge(tokens: [
+            NSNull(),
+            ["metaType": "SUPPLIED", "token": NSNull()],
+            tokenWithMetaType("SUPPLIED", address: "0xeth", balance: "1.0", balanceUSD: 2000.0, symbol: "ETH")
+        ])
+        defer { ZapperMockURLProtocol.reset() }
+        ZapperMockURLProtocol.requestHandler = { _ in
+            try (jsonData(appBalancesResponse(appEdges: [
+                NSNull(),
+                ["node": NSNull()],
+                appBalanceEdge(positionEdges: [
+                    NSNull(),
+                    ["node": NSNull()],
+                    validContractEdge,
+                    appTokenPositionEdge()
+                ])
+            ])), 200)
+        }
+
+        let provider = makeProvider(session: session)
+        let results = try await provider.fetchDeFiPositions(context: makeSyncContext())
+
+        #expect(results.count == 2)
+        #expect(results.flatMap(\.tokens).map(\.symbol) == ["ETH", "aEthUSDC"])
     }
 
     @Test
@@ -479,3 +526,5 @@ struct ZapperProviderLiveTests {
         _ = try await provider.fetchDeFiPositions(context: context)
     }
 }
+
+// swiftlint:disable:this file_length
